Checks changed, wall interaction rewrite
Changed the checks from a simple collider check, to a box area check to the final (so far) idea to use multiple raycasts to detect collisions with the ground, wall, and ceiling.

Rewriting the wall interaction script to remove spaghetti if statements and clean the code.  (also removing a couple of bugs).

Note; the wall jumping has yet to be reimplemented