Anim events for player. Some movement and knockback updates
Added an animation to state machine handler to the player to make the attack duration rely on the animation rather than a hardcoded value (for kick only so fat).  Also added an animation event for the check for hitting enemies, replacing the current (ugly) check that happens when the particular state is entered.

Movement can now be muted, e.g. during an attack the player has to stay still.

Knockback stats for the enemy are now in the enemys data rather than hardcoded